pyproject.toml
[project]
name = "yolo-auto-training"
version = "0.1.0"
description = "AI-driven automated YOLO training and deployment system"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "ultralytics>=8.0.0",
    "ray[tune]>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "crewai>=0.50.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.24.0",
    # Dataset discovery
    "requests>=2.31.0",
    "kaggle>=1.6.0",
    "datasets>=2.18.0",
    "huggingface_hub>=0.21.0",
    # Authentication
    "PyJWT>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
mlops = [
    "mlflow>=2.0.0",
    "prometheus-client>=0.19.0",
    "python-json-logger>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
